fix printing issue on wsl
Absolutely crazy bug that only manifested on WSL.

The way that printing works is there is a server that mediates
communication between the three main componnets. The server (wpexc), the
client (wp), the queue manager (wpp) and the job processor (wpp7c). They
all talk to each other over a fifo, and send each other status messages
and instructions.

The messages they send are a bit like network packets, with headers,
source and destination addresses, options and so on.

Except...what if there are two users trying to use the printer? They
solve this by each packet having a source uid so then the queue manager
can mediate access to a shared printer.

What was happening was that wp is so old that it was using a way of
checking the uid that was limited to very small numbers. The user ids
didn't match on the packet, so the queue manager canceled the jobs.

Okay - so why did this only manifest on WSL? Because Disk I/O is sloooow
on WSL, printing is so fast on native linux that the job finished
before the queue manager could cancel it :-D