rewrite mapkey() for #29, check for zeroth chars of extended sets.
raw2scan() converts the bytes read on stdin into a keycode, then
mapkey() converts that keycode into either a character or event. The way
that mapkey determines the type of scancode is checking if the
least significant byte is zero.

That usually works, but extended characters just set the upper nibble
to 0xf, and they *can* have a zero LSB, so the zeroth entry of extended
sets could not be entered from the keyboard. This is a rewrite of
mapkey() to fix that bug, however I don't 100% understand all the logic
or magic values - I've just translated them as best I can.

It *seems* to work, but requires more testing.