Add a simple CUA layout.
If you've never used WordPerfect before the key bindings might be
unfamiliar. This layout has a few common CUA bindings such as Ctrl+V for
paste, Ctrl+C for Copy and Ctrl+A for Select All.

This might make getting started easier. I added a paragraph about it to
the Getting Started wiki.

https://github.com/taviso/wpunix/wiki/Getting-Started#keyboard-layout